requirements/app.txt
galaxy.plugin.api==0.67.0
requests==2.25.0